// src/persona/entities/inversionista.dto.ts
export class RegistrarInversionista {

  //representante legal
  nrodocumento: string;
  nombre: string;
  apellido: string;
  direccion: string;
  telefono: string;
  provincia: string;
  departamento: string;
  distrito: string;

  //usuario
  correoUsuario: string;
  contrasenhaUsuario: string;
}